Separated MEF wrappers and helper class into separate files
